更新 parse-changelog 工作流 #667
