+++ app/changelogs/v0.32-sillot.md
@@ -1,0 +1,38 @@
+# Sillot v0.32
+
+
+# ❤️ 欢迎共建汐洛 694357845@qq.com
+# 🚧 汐洛仅用于开发者测试，不要用来存储重要数据！
+
+🚢 [Docker image](https://hub.docker.com/r/soltus/sillot/tags?page=1&ordering=last_updated)   📦 [Chromium Browser Extension](https://github.com/K-Sillot/Sillot-Be/releases)
+
+<span>
+<img src="https://img.shields.io/badge/Chromium 94+-black?logo=Google Chrome&logoColor=white" alt="" title=""/><img src="https://img.shields.io/badge/Windows 10+-black?logo=Windows 11" title=""/><img src="https://img.shields.io/badge/Docker-black?logo=docker" title=""/><img src="https://img.shields.io/badge/Android 12+-black?logo=android" title=""/>
+</span>
+
+---
+
+## [@Sillot](https://github.com/Hi-Windom/Sillot)
+
+### 文档相关 | Document
+
+* [更新文档](https://github.com/Hi-Windom/Sillot/issues/664)
+
+### VSCode 扩展相关 | VSCE
+
+* [汐洛VSCode插件 sofill `formatDate` 代码片段](https://github.com/Hi-Windom/Sillot/issues/666)
+
+## [@SiYuan](https://github.com/siyuan-note/siyuan)
+
+### 改进功能 | Enhancement
+
+* [Supports rendering the row number column when exporting database](https://github.com/siyuan-note/siyuan/issues/11192)
+* [When multiple files are selected in the file tree, "Open" is added to the right-click menu](https://github.com/siyuan-note/siyuan/issues/11150)
+
+### 修复错误 | Bugfix
+
+* [PDF cannot be jumped after annotation](https://github.com/siyuan-note/siyuan/issues/11197)
+* [Notebook settings doc save path does not override global settings in some cases](https://github.com/siyuan-note/siyuan/issues/11196)
+* [`matchKeymap` should not use ipcRenderer in browser](https://github.com/siyuan-note/siyuan/pull/11195)
+* [Shift+↓ cannot expand the selection of the folded list](https://github.com/siyuan-note/siyuan/issues/11194)
+
