+++ app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "sillot",
-  "version": "0.32.1100",
+  "version": "0.32.1200",
   "syv": "3.0.13",
   "sypv": "[]",
   "description": "Build Your Eternal Digital Garden",
