+++ package.cli.mjs
@@ -4,12 +4,10 @@ import inquirer from "inquirer";
 // import handlebars from "handlebars";
 // import chalk from "chalk";
 import ora from "ora";
-import fs from "fs";
-import path from "path";
 import { spawn } from "node:child_process";
 const shell = (await import("shelljs")).default;
 const iconv = (await import("iconv-lite")).default;
-import pkg from "./package.json" assert { type: "json" };
+import pkg from "./app/package.json" assert { type: "json" };
 let V = pkg.version.split(".");
 V = `v${V.at(0)}.${V.at(1)}`;
 const SYV = `v${pkg.syv}`;
